[Cart] Extract cart pickup action
