+++ src/Resources/config/services/actions/cart.xml
@@ -9,5 +9,13 @@
             <argument type="service" id="fos_rest.view_handler" />
             <argument type="service" id="sylius.shop_api_plugin.factory.cart_view_factory" />
         </service>
+
+        <service id="sylius.shop_api_plugin.controller.cart.pickup_action"
+                 class="Sylius\ShopApiPlugin\Controller\Cart\PickupAction"
+        >
+            <argument type="service" id="sylius.repository.order" />
+            <argument type="service" id="fos_rest.view_handler" />
+            <argument type="service" id="tactician.commandbus" />
+        </service>
     </services>
 </container>
