+++ tests/Controller/CartPickupApiTest.php
@@ -1,0 +1,61 @@
+<?php
+
+namespace Tests\Sylius\ShopApiPlugin\Controller;
+
+use Lakion\ApiTestCase\JsonApiTestCase;
+use League\Tactician\CommandBus;
+use Sylius\ShopApiPlugin\Command\PickupCart;
+use Symfony\Component\HttpFoundation\Response;
+
+final class CartPickupApiTest extends JsonApiTestCase
+{
+    private static $acceptAndContentTypeHeader = ['CONTENT_TYPE' => 'application/json', 'ACCEPT' => 'application/json'];
+
+    /**
+     * @test
+     */
+    public function it_creates_a_new_cart()
+    {
+        $this->loadFixturesFromFile('shop.yml');
+
+        $data =
+<<<EOT
+        {
+            "channel": "WEB_GB"
+        }
+EOT;
+
+        $this->client->request('POST', '/shop-api/carts/SDAOSLEFNWU35H3QLI5325', [], [], static::$acceptAndContentTypeHeader, $data);
+
+        $response = $this->client->getResponse();
+
+        $this->assertResponseCode($response, Response::HTTP_CREATED);
+    }
+
+    /**
+     * @test
+     */
+    public function it_does_not_allow_to_create_new_cart_if_token_is_already_used()
+    {
+        $this->loadFixturesFromFile('shop.yml');
+
+        $token = 'SDAOSLEFNWU35H3QLI5325';
+
+        /** @var CommandBus $bus */
+        $bus = $this->get('tactician.commandbus');
+        $bus->handle(new PickupCart($token, 'WEB_GB'));
+
+        $data =
+<<<EOT
+        {
+            "channel": "WEB_GB"
+        }
+EOT;
+
+        $this->client->request('POST', '/shop-api/carts/SDAOSLEFNWU35H3QLI5325', [], [], static::$acceptAndContentTypeHeader, $data);
+
+        $response = $this->client->getResponse();
+
+        $this->assertResponse($response, 'cart/token_already_used_response', Response::HTTP_BAD_REQUEST);
+    }
+}
